+++ app/Orm/Castle/Castle.php
@@ -99,7 +99,7 @@ class Castle extends \Nextras\Orm\Entity\Entity {
       $multiplier = ($this->level - 1) * 10 / 100 + 1;
     }
     $basePrice = (int) (self::BASE_REPAIR_PRICE * $multiplier * (100 - $this->hp));
-    return (int) ($basePrice - $this->eventsModel->calculateRepairingDiscount($basePrice));
+    return $basePrice - $this->eventsModel->calculateRepairingDiscount($basePrice);
   }
   
   protected function getterRepairPriceT(): string {
